ZeroInflatedDistributions

ZeroInflatedDistributions.IdentityLink

ZeroInflatedDistributions.LogitLogLink

ZeroInflatedDistributions.PoissonLink

ZeroInflatedDistributions.ZeroInflatedDistribution

ZeroInflatedDistributions.encprob

ZeroInflatedDistributions.posrate

ZeroInflatedDistributions.IdentityLink — Type

IdentityLink <: AbstractZeroInflatedLink

A simple link function where the first process is the probability of encounter and the second is expected non-zero observation.

ZeroInflatedDistributions.LogitLogLink — Type

LogitLogLink <: AbstractZeroInflatedLink

A link function where for observation y modeled by linear predictors p₁ and p₂,

p(y > 0) = logit⁻¹(p₁)

𝔼[y ∣ y > 0] = exp(p₂) - b

ZeroInflatedDistributions.PoissonLink — Type

PoissonLink <: AbstractZeroInflatedLink

The Poisson link function of Thorson (2018), where for observation y modeled by linear predictors n (log numbers-density) and w (log weight-per-group),

p(y > 0) = 1 - exp(-a exp(n))

𝔼[y ∣ y > 0] = exp(n) exp(w) / p(y > 0)

where a is an optional offset (e.g. area swept).

Encounter probability uses the complementary log-log link. This link approximates a compound Poisson-gamma (Tweedie with 1 < p < 2) distribution. This implies that number of groups sampled is drawn from a Poisson distribution with mean exp(n) and each group has expected weight exp(w).

ZeroInflatedDistributions.ZeroInflatedDistribution — Type

ZeroInflatedDistribution{Te, Tp} <: Distributions.Sampleable{Univariate,Continuous}

Inner constructor: - encdist::Bernoulli : Bernoulli distribution with probability of encounter - posdist::Distributions.ContinuousUnivariateDistribution : Continuous distribution with distribution of observations conditional on an encounter

Outer constructor: - zil::ZeroInflatedLink : Link function - posdist::Type{ContinuousUnivariateDistribution} : Type of the distribution of observations conditional on an encounter - p1 : First linear predictor - p2 : Second linear predictor - disp : Positive observation dispersion parameter (standard deviation) - biascorrect::Bool : For log-normal positive observations, parameterize by the mean rather than the median

A distribution with positive probability of a zero observation as well as a continuous response. Generally determined by two processes and a relevant link function. Outer constructors are currently implemented for the following distributions:

- LogNormal
- Gamma
- InverseGamma
- InverseGaussian

The dispersion parameter is the usual standard deviation on the log scale for log-normal, standard deviation for the gamma and inverse gamma, and the shape parameter for the inverse Gaussian distribution.

ZeroInflatedDistributions.encprob — Function

encprob(link::AbstractZeroInflatedLink, proc1, proc2)

Probability of encounter (non-zero observation) given linear predictors proc1 and proc2. Depending on the link, proc2 may not be required as an argument.

ZeroInflatedDistributions.posrate — Function

posrate(link::AbstractZeroInflatedLink, proc1, proc2)

Expected positive observation conditional on an encounter given linear predictors proc1 and proc2. Depending on the link, proc1 may be dropped as an argument.